requirements.txt
paramiko>=3.4.0
rich>=13.7.0
pydantic>=2.5.0
pyyaml>=6.0.1
